+++ test/MobilithekRequestTest.php
@@ -6,7 +6,7 @@ namespace OpenMapsight\pulpdatexfuel\dev\test;
 
 use OpenMapsight\pulp\SrcHttpHandler;
 use OpenMapsight\PulpDatexFuel;
-use OpenMapsight\pulpdatexfuel\MobilithekRequest;
+use OpenMapsight\PulpMobilithek;
 use PHPUnit\Framework\TestCase;
 use ReflectionObject;
 use RuntimeException;
@@ -22,18 +22,16 @@ class MobilithekRequestTest extends TestCase
             'Wed, 01 Jan 2025 00:00:00 GMT'
         );
 
-        $this->assertSame(PulpDatexFuel::SUBSCRIPTION_URL, MobilithekRequest::SUBSCRIPTION_URL);
+        $this->assertSame(PulpDatexFuel::SUBSCRIPTION_URL, PulpMobilithek::SUBSCRIPTION_URL);
         $this->assertSame('https://mobilithek.info:8443/mobilithek/api/v1.0/subscription', PulpDatexFuel::SUBSCRIPTION_URL);
         $this->assertSame('gzip', $options['headers']['Accept-Encoding']);
         $this->assertSame('Wed, 01 Jan 2025 00:00:00 GMT', $options['headers']['If-Modified-Since']);
-        $this->assertSame('caller-subscription-id', $options['query']['subscriptionID']);
+        $this->assertSame(['subscriptionID' => 'caller-subscription-id'], $options['query']);
         $this->assertSame('/tmp/client.p12', $options['curl'][CURLOPT_SSLCERT]);
         $this->assertSame('caller-password', $options['curl'][CURLOPT_SSLCERTPASSWD]);
         $this->assertSame('P12', $options['curl'][CURLOPT_SSLCERTTYPE]);
         $this->assertTrue($options['decode_content']);
         $this->assertFalse($options['http_errors']);
-        $this->assertStringNotContainsString('1030105573172883456', json_encode($options, JSON_THROW_ON_ERROR));
-        $this->assertStringNotContainsString('1030117702173069312', json_encode($options, JSON_THROW_ON_ERROR));
     }
 
     public function testSrcMobilithekReturnsConfiguredSrcHttpHandler(): void